Refactored overview component into smaller ones
